Move styleConstants to styles dir
