+++ src/app/components/Typography/HeadingText/styles.js
@@ -1,6 +1,6 @@
 import css from 'styled-jsx/css'; // eslint-disable-line
 
-import { colors } from '../../../static/styleConstants';
+import { colors } from '../../../static/styles/styleConstants';
 
 const styles = css`
   h2 {
