+++ src/app/static/styles/helpers/index.js
@@ -1,6 +1,6 @@
 import css from 'styled-jsx/css';
 
-import { rhythm } from '../../styleConstants';
+import { rhythm } from '../styleConstants';
 
 const styles = css.global`
   .flex {
