+++ src/app/static/styles/styleConstants/index.js
